requirements/develop.txt
-r base.txt
django-debug-toolbar
model_mommy
ipdb